==> scripts/8_Figure3_Network_map_top90perc.Rmd <==
---
title: "Untitled"
author: "Caitie"
date: '2022-07-21'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
#install.packages("tidytuesdayR")
library(ggnetwork)
#devtools::install_github("yixuan/sysfonts")
library(sysfonts)
library(showtext)
library(raster)
library(grid)
library(gridExtra)
library(tidyverse)
library(sp)
library(rworldmap)
library(sf)
library(rgeos)
library(countrycode)
library(here)
```
# Load data

```{r}

combos<-read.csv(here("output_data/Country_redundant_trade_combinations.csv")) %>% 
  dplyr::select(x ,y, redun_like_tonnes) %>% 
  rename(ExportISO3 = x,
         ImportISO3 = y) %>% 
  rename(bad_redun = redun_like_tonnes) %>% 
  mutate(redun_like_tonnes = bad_redun/2)
```


# Import Rubik font from Google

```{r}
font_add_google("Rubik", "Rubik")
showtext_auto()
```


# Get country lat/long centroids

```{r}
# get world map
wmap <- getMap(resolution="high")

# get centroids
centroids <- gCentroid(wmap, byid=TRUE)

# get a data.frame with centroids
df <- as.data.frame(centroids)

# convert row names into first column
df <- tibble::rownames_to_column(df, "Country")

# Add in ISO3 codes
df <- df %>% 
  mutate(iso3 = countrycode(Country, "country.name", "iso3c")) %>% 
  filter(is.na(iso3) == F,
         !iso3 == "ATA") %>% # remove countries without iso3 for now
  filter(!Country == "US Naval Base Guantanamo Bay")
```

# pivot data into matrix format
```{r}
mat <- combos %>%
  pivot_wider(id_cols = ExportISO3, names_from = ImportISO3, values_from = redun_like_tonnes) %>%
  column_to_rownames(var="ExportISO3")
```

```{r}
# replace NA with 0
mat[is.na(mat)] <- 0

# create matrix
mat <- as.matrix(mat)
```


# long/lat join df for from_country
```{r}
country_join_a <- df %>%
  dplyr::select(ExportISO3 = iso3, x, y)
```

# long/lat join df for to_country
```{r}
country_join_b <- df %>%
  dplyr::select(ImportISO3 = iso3,x,y) %>% 
  rename(xend = x,
         yend = y)
```

```{r}

countries<-unique(c(unique(combos$ExportISO3), unique(combos$ImportISO3)))

cumsum<-combos %>% 
  arrange(desc(redun_like_tonnes)) %>% 
  ungroup() %>% 
  mutate(cumsum = cumsum(redun_like_tonnes),
         prop_cumsum = cumsum/sum(combos$redun_like_tonnes, na.rm = T)) %>% 
  filter(prop_cumsum<=0.9)

sub_countries<-unique(c(unique(cumsum$ImportISO3), unique(cumsum$ExportISO3)))

# create df for plotting with both sets of long/lat values
plot_df <- data.frame(ExportISO3 = rownames(mat)) %>%
  full_join(cumsum, by = "ExportISO3") %>%
  full_join(country_join_a, by = c("ExportISO3")) %>%
  full_join(country_join_b, by = c("ImportISO3")) %>%
  arrange(redun_like_tonnes) %>% 
  mutate(from_name = countrycode(ExportISO3, "iso3c", "country.name"),
         label = ifelse(ExportISO3 %in% countries, ExportISO3, NA),
         label_90 = ifelse(ExportISO3 %in% sub_countries, ExportISO3, NA),
         label_country = countrycode(label_90, "iso3c", "country.name"))

plot_df[73, 13]<-NA



```

# create top 5 section:

```{r}
top_table <- combos %>%
  arrange(-redun_like_tonnes) %>%
  head(10) 

top_table$rank<-1:nrow(top_table)
```

# create simple world map

```{r}
map <- map_data("world") %>% 
  filter(!region == "Antarctica") %>% 
  mutate(iso3 = countrycode(region, "country.name", "iso3c")) %>% 
  dplyr::select(x = long, y = lat, group, id = subregion, iso3) %>%
  mutate(xend = 0,
         yend = 0)
```


# build plot


```{r}
ggplot(plot_df, aes(x = x, y = y, xend = xend, yend = yend)) +
  geom_polygon(data = map, aes(group = group), fill = "White", colour = "darkgray") +
  geom_edges(aes(color = redun_like_tonnes/1000), curvature = -0.3, size = 0.7) +
  geom_nodetext(aes(label = str_wrap(toupper(label_90), 8)), size = 10, color = "black",
                fontface = "bold", repel = TRUE) +
   guides(size = "none", alpha = "none") +
  labs(color = "Redundant trade\n(thousand tonnes)", caption = "") +
  scale_color_viridis_b(direction = -1) +
  theme_blank() +
    theme(
    legend.position = "bottom",
   # axis.title.x = element_text(),
    legend.key.width = unit(2, "cm"),
    legend.key.height = unit(0.3, "cm"),
    plot.title = element_text(hjust = 0.5, face = "bold"),
    #plot.background = element_rect(fill = "darkgray", color = "darkgray"),
    #panel.background = element_rect(fill = "darkgray", color = "darkgray"),
   # text = element_text(family = "Rubik", color = "black"),
    #legend.background = element_rect(fill = "darkgray"),
    legend.margin = margin(unit = "cm") 
  )
```

```{r}
ggsave(here("figures/Figure3a_Redun_trade_network_map.png"), dpi = 300, width = 16, height = 8)
```


```{r}
continents<-read.csv(here("raw_data/Countries_by_continent.csv"))

europe <- map_data("world") %>% 
  full_join(., continents, by = c("region" = "Country")) %>% 
  filter(Continent == "Europe") %>% 
  mutate(iso3 = countrycode(region, "country.name", "iso3c")) %>% 
  dplyr::select(x = long, y = lat, group, id = subregion, iso3) %>%
  mutate(xend = 0,
         yend = 0)
```

```{r}
sub_plot_df <- plot_df %>% 
  full_join(., continents, by = c("ExportISO3" = "ISO3")) %>% 
  filter(Continent == "Europe")
```


```{r}
ggplot(sub_plot_df, aes(x = x, y = y, xend = xend, yend = yend)) +
  geom_polygon(data = europe, aes(group = group), fill = "White", colour = "darkgray") +
  geom_edges(aes(color = redun_like_tonnes/1000), curvature = -0.2) +#, arrow = arrow(length = unit(6, "pt"), type = "closed")) +
  #geom_edges(arrow = arrow(length = unit(4, "pt"), type = "closed")) +
  #geom_nodes(size = 2, color = "#1c192c", alpha = 0.8) +
  geom_nodetext(aes(label = str_wrap(toupper(label_90), 8)), size = 3.5, color = "black",
                fontface = "bold") +
   guides(size = "none", alpha = "none") +
  labs(color = "Redundant trade\n(thousand tonnes)", caption = "") +
  scale_color_viridis_b() +
  theme_blank() +
    theme(
    legend.position = "bottom",
   # axis.title.x = element_text(),
    legend.key.width = unit(2, "cm"),
    legend.key.height = unit(0.3, "cm"),
    plot.title = element_text(hjust = 0.5, face = "bold"),
    #plot.background = element_rect(fill = "darkgray", color = "darkgray"),
    #panel.background = element_rect(fill = "darkgray", color = "darkgray"),
   # text = element_text(family = "Rubik", color = "black"),
    #legend.background = element_rect(fill = "darkgray"),
    legend.margin = margin(unit = "cm")
  ) #+
  #scale_x_continuous(limits = c(-125, -67))+
  #      scale_y_continuous(limits = c(25, 50))
```

## CHORD DIAGRAM

```{r}
library(rnaturalearth)
library(rnaturalearthdata)
library(circlize)
#devtools::install_github("https://github.com/mattflor/chorddiag")
library(chorddiag)
library(randomcoloR)

```

# Top 90% of redundant trade chord diagram

```{r}
test<-cumsum %>% 
  #mutate(from = countrycode(ExportISO3, "iso3c", "country.name"),
  #       to = countrycode(ImportISO3, "iso3c", "country.name"),
   #      value = redun_like_tonnes/1000) %>% 
  rename(from = ExportISO3,
         to = ImportISO3,
         value = redun_like_tonnes) %>% 
  mutate(value = value/1000) %>% 
  dplyr::select(from, to, value)
```



Try changing things based on this page:

https://jokergoo.github.io/circlize_book/book/the-chorddiagram-function.html#basic-usage-of-making-chord-diagram
```{r}
pal <- brewer.pal(9, "YlOrRd")

circos.par(gap.after = 6,
           start.degree = 85, 
           clock.wise = FALSE)#,
           #circle.margin = 0.05)

chordDiagram(test,  grid.col = "gray", col = pal, transparency = 0.2, directional = 1,
             annotationTrack = "grid", preAllocateTracks = 1,
             link.border = "darkgray")
circos.trackPlotRegion(track.index = 1, panel.fun = function(x, y) {
  xlim = get.cell.meta.data("xlim")
  ylim = get.cell.meta.data("ylim")
  sector.name = get.cell.meta.data("sector.index")
  circos.text(mean(xlim), ylim[1] + .25, sector.name, facing = "clockwise", niceFacing = TRUE, adj = c(0, 0.7))
  circos.axis(h = "top", labels.cex = 0.01, sector.index = sector.name, track.index = 2)
}, bg.border = NA)

circos.clear()
```
```{r}
png(here("figures/Figure3b_Circle_plot_top90perc.png"))

circos.par(gap.after = 6,
           start.degree = 85, 
           clock.wise = FALSE)

#change col = "gray" to col = pal to get color palette back

chordDiagram(test,  grid.col = "gray", col = "gray", transparency = 0.2, directional = 1,
             annotationTrack = "grid", preAllocateTracks = 1,
             link.border = "darkgray")
circos.trackPlotRegion(track.index = 1, panel.fun = function(x, y) {
  xlim = get.cell.meta.data("xlim")
  ylim = get.cell.meta.data("ylim")
  sector.name = get.cell.meta.data("sector.index")
  circos.text(mean(xlim), ylim[1] + .25, sector.name, facing = "clockwise", niceFacing = TRUE, adj = c(0, 0.5))
  circos.axis(h = "top", labels.cex = 0.01, sector.index = sector.name, track.index = 2)
}, bg.border = NA)

dev.off()
```

```{r}
lgd_chr = Legend(title = "Chromatin States", at = names(state_col), 
        legend_gp = gpar(fill = state_col))))
    lgd_mmeth = Legend(title = "Mean Methylation", at = seq(0.1, 0.9, 0.2), col_fun = col_fun)
    lgd_mdmeth = Legend(title = "Mean Difference", col_fun = col_fun2)
        
    h = dev.size()[2]
    circle_size = unit(1, "snpc")
    lgd_list = packLegend(lgd_chr, lgd_mmeth, lgd_mdmeth, max_height = unit(0.9*h, "inch"))
    draw(lgd_list, x = circle_size, just = "right") 
```
