==> scripts/9_Figure4_Species_figure.Rmd <==
---
title: "Untitled"
author: "Caitie"
date: '2022-07-22'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(ggplot2)
library(viridis)
library(circlize)

options(scipen = 999)
```

```{r}
continents<-read.csv(here("raw_data/Countries_by_continent.csv"))
```

```{r}
#thr_files<-list.files("/home/kuempel/IUCN_Ohara/spp_risk_dists/_data/", pattern = "iucn_risk_current", full.names = T)

#thr_dat<-do.call(rbind, lapply(thr_files, read.csv)) %>% 
#  dplyr::select(sciname, cat) %>% 
#  distinct()
```


```{r}
dat<-read.csv(here("output_data/Country_redundant_species_results.csv"))

spp_dat<-dat %>% 
  group_by(TaxonName, CommonName) %>% 
  summarise(total_like = sum(like_tonnes, na.rm = T),
            total_redun = sum(redun_tonnes, na.rm = T))

spp_dat<-spp_dat %>% 
  arrange(desc(total_redun))

write.csv(spp_dat, here("output_data/Redundant_species_list.csv"))

top_spp<-spp_dat[1:4,]


spp_dat<-spp_dat %>% 
  mutate(prop = total_redun/sum(spp_dat$total_redun, na.rm = T)) %>% 
  filter(!is.na(total_redun) == T) %>% 
  arrange(desc(prop)) %>% 
  ungroup() %>% 
  mutate(cumsum_prop = cumsum(prop))

top_90perc<-spp_dat %>% 
  filter(cumsum_prop<=0.9)

#thr_spp<-left_join(top_spp, thr_dat, by = c("TaxonName" = "sciname")) %>% 
 # filter(!is.na(redun_tonnes) == T) %>% 
#    mutate(cat = ifelse(is.na(cat) == T, "Unknown", as.character(cat)))
```

```{r}
dat_year<-dat %>% 
  group_by(Year, CommonName, TaxonName) %>% 
  summarise(total_redun = sum(redun_tonnes, na.rm = T)) %>% 
  filter(CommonName %in% top_90perc$CommonName) %>% 
  mutate(Sci_name = paste0("(",TaxonName,")"),
         label = paste0(CommonName, " ", Sci_name),
         label = ifelse(label == "Atlantic cod (Gadus morhua)", "Atlantic cod (Gadus morhua)**",
         ifelse(label =="Atlantic horse mackerel (Trachurus trachurus)", "Atlantic horse mackerel (Trachurus trachurus)**", as.character(label))))
```
Atlantic cod is Vulnerable
Atlantic horse mackerel is Vulnerable

```{r}
pal <- brewer.pal(9, "YlGnBu")

ggplot(data = dat_year, mapping = aes(x = Year,y = reorder(label, total_redun),fill = total_redun/10000)) +
  geom_tile() + 
  #scale_fill_viridis(direction =-1)+
  scale_fill_gradientn(colours = pal) +
  theme_minimal() +
  ylab("Species") +
  labs(fill = "Tonnes redundant\ntrade (10,000)") +
  theme(text = element_text(size = 10)) +
  theme(legend.position = "bottom")
```
```{r}
ggsave(here("figures/Figure4a_Species_year_heatmap.png"), dpi = 300, width = 6, height = 5)
```



```{r}
iso3_dat<-dat %>% 
  filter(dat$TaxonName %in% top_spp$TaxonName)

iso3_dat2<- iso3_dat %>% 
  group_by(c1, c2, TaxonName, CommonName) %>% 
  summarise(total_like = sum(like_tonnes, na.rm = T),
            total_redun = sum(redun_tonnes, na.rm = T)) %>% 
  left_join(., continents, by = c("c1" = "ISO3"))
```

```{r}
sp1<-iso3_dat2 %>% 
  filter(CommonName == "Skipjack tuna") %>% 
  ungroup() %>% 
  dplyr::select(c1,c2, total_redun)
sp2<-iso3_dat2 %>% 
  filter(CommonName == "Atlantic herring")%>% 
  ungroup() %>% 
  dplyr::select(c1,c2, total_redun)
sp3<-iso3_dat2 %>% 
  filter(CommonName == "Atlantic cod")%>% 
  ungroup() %>% 
  dplyr::select(c1,c2, total_redun)
sp4<-iso3_dat2 %>% 
  filter(CommonName == "Atlantic mackerel")%>% 
  ungroup() %>% 
  dplyr::select(c1,c2, total_redun)
```

```{r}
png(here("figures/Figure4b_Circle_plot_skipjack_tuna.png"))
pal <- brewer.pal(9, "YlOrRd")

circos.par(gap.after = 6,
           start.degree = 85, 
           clock.wise = FALSE)#,
           #circle.margin = 0.05)

chordDiagram(sp1,  grid.col = "gray", col = "gray", transparency = 0.2, directional = 1,
             annotationTrack = "grid", preAllocateTracks = 1,
             link.border = "darkgray")
circos.trackPlotRegion(track.index = 1, panel.fun = function(x, y) {
  xlim = get.cell.meta.data("xlim")
  ylim = get.cell.meta.data("ylim")
  sector.name = get.cell.meta.data("sector.index")
  circos.text(mean(xlim), ylim[1] + .25, sector.name, facing = "clockwise", niceFacing = TRUE, adj = c(0, 0.7))
  circos.axis(h = "top", labels.cex = 0.01, sector.index = sector.name, track.index = 2)
}, bg.border = NA) +
  title("Skipjack tuna (Katsuwonus pelamis)")

circos.clear()

dev.off()
```

```{r}
png(here("figures/Figure4c_Circle_plot_Atlantic_herring.png"))
pal <- brewer.pal(9, "YlOrRd")

circos.par(gap.after = 6,
           start.degree = 85, 
           clock.wise = FALSE)#,
           #circle.margin = 0.05)

chordDiagram(sp2,  grid.col = "gray", col = "gray", transparency = 0.2, directional = 1,
             annotationTrack = "grid", preAllocateTracks = 1,
             link.border = "darkgray")
circos.trackPlotRegion(track.index = 1, panel.fun = function(x, y) {
  xlim = get.cell.meta.data("xlim")
  ylim = get.cell.meta.data("ylim")
  sector.name = get.cell.meta.data("sector.index")
  circos.text(mean(xlim), ylim[1] + .25, sector.name, facing = "clockwise", niceFacing = TRUE, adj = c(0, 0.7))
  circos.axis(h = "top", labels.cex = 0.01, sector.index = sector.name, track.index = 2)
}, bg.border = NA) +
  title("Atlantic herring (Clupea harengus)")

circos.clear()
dev.off()
```
```{r}
png(here("figures/Figure4d_Circle_plot_Atlantic_cod.png"))
pal <- brewer.pal(9, "YlOrRd")

circos.par(gap.after = 6,
           start.degree = 85, 
           clock.wise = FALSE)#,
           #circle.margin = 0.05)

chordDiagram(sp3,  grid.col = "gray", col = "gray", transparency = 0.2, directional = 1,
             annotationTrack = "grid", preAllocateTracks = 1,
             link.border = "darkgray")
circos.trackPlotRegion(track.index = 1, panel.fun = function(x, y) {
  xlim = get.cell.meta.data("xlim")
  ylim = get.cell.meta.data("ylim")
  sector.name = get.cell.meta.data("sector.index")
  circos.text(mean(xlim), ylim[1] + .25, sector.name, facing = "clockwise", niceFacing = TRUE, adj = c(0, 0.7))
  circos.axis(h = "top", labels.cex = 0.01, sector.index = sector.name, track.index = 2)
}, bg.border = NA) +
  title("Atlantic cod (Gadus morhua)")

circos.clear()

dev.off()
```

```{r}
png(here("figures/Figure4e_Circle_plot_Atlantic_mackerel.png"))
pal <- brewer.pal(9, "YlOrRd")

circos.par(gap.after = 6,
           start.degree = 85, 
           clock.wise = FALSE)#,
           #circle.margin = 0.05)

chordDiagram(sp4,  grid.col = "gray", col = "gray", transparency = 0.2, directional = 1,
             annotationTrack = "grid", preAllocateTracks = 1,
             link.border = "darkgray")
circos.trackPlotRegion(track.index = 1, panel.fun = function(x, y) {
  xlim = get.cell.meta.data("xlim")
  ylim = get.cell.meta.data("ylim")
  sector.name = get.cell.meta.data("sector.index")
  circos.text(mean(xlim), ylim[1] + .25, sector.name, facing = "clockwise", niceFacing = TRUE, adj = c(0, 0.7))
  circos.axis(h = "top", labels.cex = 0.01, sector.index = sector.name, track.index = 2)
}, bg.border = NA) +
  title("Atlantic mackerel (Scomber scombrus)")

circos.clear()

dev.off()
```
